refactor(client): topnavigator hooks 리팩토링을 진행합니다
