+++ .eslintrc.json
@@ -14,7 +14,7 @@
     "plugin:@typescript-eslint/recommended"
   ],
   "rules": {
-    "import/prefer-default-export": 1, // 0 -> 0ff / 1 -> on (default)
+    "import/prefer-default-export": 0, // 0 -> 0ff / 1 -> on (default)
     "jsx-a11y/no-noninteractive-element-interactions": 0,
     "prettier/prettier": 0,
     "import/extensions": 0,
